+++ CHANGELOG.rst
@@ -1,9 +1,14 @@
-Version 7.2.0
+Version 8.0.0
 -------------
 
 **Added**
 
 - ``bank()`` method for ``Finance()``.
+- ``default_country`` for ``Address()``, which always returns the country associated with the current locale (i.e ``United States`` for ``en``, ``Россия`` for ``ru``).
+
+**Removed**:
+
+- Removed parameter ``allow_random`` for ``country()``. Now method returns random country by default.
 
 Version 7.1.0
 -------------